requirements.txt
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
Deprecated==1.2.14
Flask==2.3.2
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gcloud==0.18.3
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.2
oauth2client==4.1.3
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pyparsing==3.0.9
Pyrebase4==4.7.0
python-jwt==4.0.0
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.15
typing-extensions==4.6.2
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
